chore: inprove vulnerability report, and add parameter for quick parsing already generated target files

+++ scripts/generateAndCheckSBOM.js
@@ -71,9 +71,10 @@ for (let i = 2, l = process.argv.length; i < l; i++) {
     case '--enable-full-owasp': cmd.useFullOWASP = true; break;
     case '--version': cmd.version = process.argv[++i]; break;
     case '--compare': cmd.org = process.argv[++i]; break;
+    case '--quick': cmd.quick = true; break;
     default:
       console.log(`Usage: ${path.relative('.', process.argv[1])}
-       [--useSnapshots] [--disable-bomber] [--disable-osv-scan] [--disable-owasp] [--enable-full-owasp] [--version x.x.x]`);
+       [--useSnapshots] [--disable-bomber] [--disable-osv-scan] [--disable-owasp] [--enable-full-owasp] [--version x.x.x] [--quick]`);
       process.exit(1);
   }
 }
@@ -353,7 +354,7 @@ function checkVunerabilities(vuls) {
   Object.keys(vuls).forEach(v => {
     const cves = Object.keys(vuls[v]).sort().join(', ');
     err = err || (!cveWhiteList[v] || cves !== cveWhiteList[v].sort().join(', '));
-    msg += `  - Vulnerabilities in: ${v} [${Object.keys(vuls[v]).join(', ')}]\n`;
+    msg += `  - Vulnerabilities in: ${v} [${Object.keys(vuls[v]).join(', ')}] (${[...new Set(Object.values(vuls[v]).flatMap(o => o.scanner))].join(',')})\n`;
   });
   return { err, msg };
 }
@@ -390,10 +391,11 @@ function reportLicenses(licenses) {
 function reportVulnerabilities(vuls) {
   let md = "", html = "";
   Object.keys(vuls).forEach(v => {
+    const title = o => o.title.replace(/[^\w\d]/g, ' ');
     html += `<tr><td><code>${v}</code></td><td><ul><li>${Object.keys(vuls[v]).map(o =>
-      `<a href="https://nvd.nist.gov/vuln/detail/${o}">${o}</a> <i>${vuls[v][o].title}</i> (${[...new Set(vuls[v][o].scanner)].join(',')})`).join('<li>')}</ul></td></tr>\n`;
+      `<a href="https://nvd.nist.gov/vuln/detail/${o}">${o}</a> <i>${title(vuls[v][o])}</i> (${[...new Set(vuls[v][o].scanner)].join(',')})`).join('<li>')}</ul></td></tr>\n`;
     md += `|\`${v}\`|<ul><li>${Object.keys(vuls[v]).map(o =>
-      `[${o}](https://nvd.nist.gov/vuln/detail/${o}) _${vuls[v][o].title}_ (${[...new Set(vuls[v][o].scanner)].join(',')})`).join('<li>')}</ul>\n`;
+      `[${o}](https://nvd.nist.gov/vuln/detail/${o}) _${title(vuls[v][o])}_ (${[...new Set(vuls[v][o].scanner)].join(',')})`).join('<li>')}</ul>\n`;
   });
   html && (html = `<table><tr><th>Package</th><th>CVEs</th>\n${html}</table>\n`)
   md && (md = "| Package | CVEs |\n|-------|--------|\n" + md);
@@ -449,29 +451,30 @@ async function main() {
   await isInstalled('mvn');
   await isInstalled('curl');
 
-  if (cmd.version) {
+  if (!cmd.quick && cmd.version) {
     await run(`mvn -ntp -N -B -DnewVersion=${cmd.version} -Psbom versions:set -q`);
   }
 
   const currVersion = cmd.version || (await run('mvn help:evaluate -N -q -DforceStdout -Dexpression=project.version', { debug: false })).stdout;
   log(`current version: ${currVersion}`);
-  await run(`./scripts/generateBoms.sh${cmd.useSnapshots ? ' --useSnapshots' :''}`, { debug: false });
-  await run('mvn -ntp -B clean install -T 1C -q -DskipTests');
+  if (!cmd.quick) {
+    await run(`./scripts/generateBoms.sh${cmd.useSnapshots ? ' --useSnapshots' :''}`, { debug: false });
+    await run('mvn -ntp -B clean install -T 1C -q -DskipTests');
+  }
 
   log(`cd ${testProject}`);
   process.chdir(testProject);
 
-  log(`cleaning package.json`);
-  fs.existsSync('package.json') && fs.unlinkSync('package.json');
-
-  await run('mvn clean package -ntp -B -Pproduction -DskipTests -q');
-  await run('mvn dependency:tree -ntp -B', { output: 'target/tree-maven.txt' });
-  await run('mvn -ntp -B org.cyclonedx:cyclonedx-maven-plugin:makeAggregateBom -q');
-  await run('npm ls --depth 6 --omit dev', { output: 'target/tree-npm.txt' });
-
-  await run('npm install');
-  await run('npm install @cyclonedx/cyclonedx-npm');
-  await run('npx @cyclonedx/cyclonedx-npm --omit dev --output-file target/bom-npm.json --output-format JSON');
+  if (!cmd.quick) {
+    fs.existsSync('package.json') && log(`cleaning package.json`) && fs.unlinkSync('package.json');
+    await run('mvn clean package -ntp -B -Pproduction -DskipTests -q');
+    await run('mvn dependency:tree -ntp -B', { output: 'target/tree-maven.txt' });
+    await run('mvn -ntp -B org.cyclonedx:cyclonedx-maven-plugin:makeAggregateBom -q');
+    await run('npm ls --depth 6 --omit dev', { output: 'target/tree-npm.txt' });
+    await run('npm install');
+    await run('npm install @cyclonedx/cyclonedx-npm');
+    await run('npx @cyclonedx/cyclonedx-npm --omit dev --output-file target/bom-npm.json --output-format JSON');
+  }
 
   log(`generating 'bom-vaadin.js'`);
   const sbom = await consolidateSBoms('target/bom.json', 'target/bom-npm.json');
@@ -481,31 +484,32 @@ async function main() {
 
   const vulnerabilities = {}
   if (cmd.useBomber) {
-    const cmdBomber = `bomber scan target/bom-vaadin.json --output json`;
-    await run(cmdBomber, { output: 'target/bomber-osv-report.json' });
-    sumarizeBomber('target/bomber-osv-report.json', vulnerabilities);
-    if (cmd.hasOssToken) {
-      await run(`${cmdBomber} --provider ossindex --username ${process.env.OSSINDEX_USER} --token ${process.env.OSSINDEX_TOKEN}`,
+    if (!cmd.quick) {
+      const cmdBomber = `bomber scan target/bom-vaadin.json --output json`;
+      await run(cmdBomber, { output: 'target/bomber-osv-report.json' });
+      cmd.hasOssToken && await run(
+        `${cmdBomber} --provider ossindex --username ${process.env.OSSINDEX_USER} --token ${process.env.OSSINDEX_TOKEN}`,
         { output: 'target/bomber-oss-report.json' });
-      sumarizeBomber('target/bomber-oss-report.json', vulnerabilities);
     }
+    sumarizeBomber('target/bomber-osv-report.json', vulnerabilities);
+    fs.existsSync('target/bomber-oss-report.json') && sumarizeBomber('target/bomber-oss-report.json', vulnerabilities);
   }
 
   if (cmd.useOSV) {
-    await run('osv-scanner --sbom=target/bom-vaadin.json --json', { output: 'target/osv-scanner-report.json', throw: false });
+    !cmd.quick && await run('osv-scanner --sbom=target/bom-vaadin.json --json', { output: 'target/osv-scanner-report.json', throw: false });
     sumarizeOSV('target/osv-scanner-report.json', vulnerabilities);
   }
 
   if (cmd.useOWASP) {
     // https://github.com/jeremylong/DependencyCheck/issues/4293
     // https://github.com/jeremylong/DependencyCheck/issues/1947
-    fs.unlinkSync('package-lock.json')
-    await run('mvn org.owasp:dependency-check-maven:check -Dformat=JSON -q', { throw: false });
+    fs.existsSync('package-lock.json') && fs.unlinkSync('package-lock.json')
+    !cmd.quick && await run('mvn org.owasp:dependency-check-maven:check -Dformat=JSON -q', { throw: false });
     sumarizeOWASP('target/dependency-check-report.json', vulnerabilities);
   }
 
   if (cmd.useFullOWASP) {
-    await run('dependency-check -f JSON -f HTML --prettyPrint --out target --scan .');
+    !cmd.quick && await run('dependency-check -f JSON -f HTML --prettyPrint --out target --scan .');
     sumarizeOWASP('target/dependency-check-report.json', vulnerabilities);
   }
